added logout route and removed login function from utils and moved it to controller
