+++ router/userRoutes.js
@@ -1,9 +1,9 @@
 import express from "express";
-import { register } from "../controllers/userController.js";
-import { login } from "../utils/jwtToken.js";
+import { logout, register, login } from "../controllers/userController.js";
 
 const router = express.Router();
 router.post("/user/register", register);
 router.post("/user/login", login);
+router.post("/user/logout", logout);
 
 export default router;
